+++ netomaton/utils.py
@@ -170,11 +170,15 @@ def plot_network(network, layout="shell", with_labels=True, node_color="#1f78b4"
 
 
 def animate_network(trajectory, save=False, interval=50, dpi=80, layout="shell",
-                    with_labels=True, node_color="b", node_size=30):
+                    with_labels=True, with_arrows=True, node_color="b", node_size=30):
     fig, ax = plt.subplots()
 
-    def update(state):
+    def update(arg):
         ax.clear()
+
+        i, state = arg
+
+        color = node_color[i] if type(node_color) == dict else node_color
 
         network = state.network
         G = nx.MultiDiGraph()
@@ -183,16 +187,22 @@ def animate_network(trajectory, save=False, interval=50, dpi=80, layout="shell",
         for edge in network.edges:
             G.add_edge(edge[0], edge[1])
 
-        if layout == "shell":
-            nx.draw_shell(G, with_labels=with_labels, node_color=node_color, node_size=node_size)
-        elif layout == "spring":
-            nx.draw_spring(G, with_labels=with_labels, node_color=node_color, node_size=node_size)
-        elif isinstance(layout, dict):
-            nx.draw(G, pos=layout, with_labels=with_labels, node_color=node_color, node_size=node_size)
-        else:
-            raise Exception("unsupported layout: %s" % layout)
+        state_layout = layout[i] if type(layout) == list or type(layout) == tuple else layout
 
-    ani = animation.FuncAnimation(fig, update, frames=trajectory, interval=interval,
+        if state_layout == "shell":
+            nx.draw_shell(G, with_labels=with_labels, node_color=color, node_size=node_size, arrows=with_arrows)
+        elif state_layout == "spring":
+            nx.draw_spring(G, with_labels=with_labels, node_color=color, node_size=node_size, arrows=with_arrows)
+        elif state_layout == "planar":
+            nx.draw_planar(G, with_labels=with_labels, node_color=color, node_size=node_size, arrows=with_arrows)
+        elif state_layout == "kamada-kawai":
+            nx.draw_kamada_kawai(G, with_labels=with_labels, node_color=color, node_size=node_size, arrows=with_arrows)
+        elif isinstance(state_layout, dict):
+            nx.draw(G, pos=state_layout, with_labels=with_labels, node_color=color, node_size=node_size, arrows=with_arrows)
+        else:
+            raise Exception("unsupported layout: %s" % state_layout)
+
+    ani = animation.FuncAnimation(fig, update, frames=list(enumerate(trajectory)), interval=interval,
                                   save_count=len(trajectory))
     if save:
         ani.save('evolved.gif', dpi=dpi, writer="imagemagick")