+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `CHANGELOG.md`
 - Added `MemoizationKey` and memoization support
+- Added demo for functional network automata
+- Added demo for evolving networks
 
 ### Changed